Move creation of installed packages back into rupm
jtar.sh is removed, removing a dependency on jshon (which is otherwise
perfectly fine software) for creating packages. rupm does not depend on
any non-POSIX tools anymore (as far as I know). The drawback of this is
that generation of packages is not as flexible anymore. This is fine,
since this is the purpose of build tools. (rupm can still package
correctly placed installed programs, just not move them around)

Another decision is quite definitive this commit. rupm will not handle
paths with newlines in them, and will probably even break horribly. If I
ever rewrite this in C or Rust or whatever (which probably won't happen,
at most a specification-correct clone) this will be fixed, but working
around that in pure POSIX sh is just not possible/worth it.