Move tarxenv.sh into rupm.sh
This means that rupm is now a single shell script. I intend to keep
it that way. It is easier to work around bootstrapping problems this
way. (Not that I have run into any of those, but it made no sense to
pollute the PATH with a program nobody will use.) I might add some
utilities in the future, but those will probably belong in a separate
package/repo/whatever.

A drawback is that we can't use parts written in shell script as the
executable in a `find(1)` command.