Add package version stub
This hardcodes the package file name convention. I think that is
preferable to non-hardcoded, as this convention allows some useful
extractions.

I am not entirely happy with a revision separator that's the same
as the version separator '.'. I might change that, but for now it
allows the `sort` invocation to be POSIX-compatible. Maybe I can find
a workaround in the future.